feat: Add cancel follow request functionality
Add a new route and controller function to handle canceling follow requests. This allows users to cancel their pending follow requests to other users. The new route is "/user/follow/request/cancel/user/follow-user-request-user-relation" and the corresponding controller function is CancelRequest(). This commit enhances the functionality of the application by providing users with the ability to cancel follow requests.