chore: Update ViewFollowing and ViewFollowers functions to use usersearchID parameter
Update the ViewFollowing and ViewFollowers functions in the relations.go controller to use the usersearchID parameter instead of the userIDObj.Hex(). This change improves the flexibility of the functions by allowing users to view the followers and following of a specific user by providing their ID as a query parameter.